Switch to full hash everywhere
Now that the dag_hash is computed using build dependencies and the
package content hash (and is equivalent to full_hash), use the new
dag_hash everywhere in spack.  We keep the definitions of the full_hash
and build_hash so we can recognize them in old spec files, but they
should no longer be stored on spec objects or used in any other way.